requirements.txt
asgiref
cryptography
sqlparse
Django
dotenv
python-decouple
mysqlclient
requests
celery
django-multiselectfield
django-admin-sortable2
aiohttp
redis
django-celery-beat
